+++ tests/TestCase/View/Cell/PhotosWidgetsCellTest.php
@@ -13,55 +13,57 @@
 namespace MeCmsInstagram\Test\TestCase\View\Cell;
 
 use Cake\Cache\Cache;
+use MeCms\TestSuite\CellTestCase;
 use MeCms\View\Helper\WidgetHelper;
 use MeCms\View\View\AppView as View;
-use MeTools\TestSuite\TestCase;
 
 /**
  * PhotosWidgetsCellTest class
  */
-class PhotosWidgetsCellTest extends TestCase
+class PhotosWidgetsCellTest extends CellTestCase
 {
     /**
-     * @var \MeCms\View\Helper\WidgetHelper
-     */
-    protected $Widget;
-
-    /**
-     * Setup the test case, backup the static object values so they can be
-     * restored. Specifically backs up the contents of Configure and paths in
-     *  App if they have not already been backed up
+     * Called before every test method
      * @return void
      */
     public function setUp()
     {
-        parent::setUp();
-
-        Cache::clearAll();
-
         $this->Widget = $this->getMockBuilder(WidgetHelper::class)
             ->setConstructorArgs([new View])
             ->setMethods(['widget'])
             ->getMock();
 
         $this->Widget->method('widget')->will($this->returnCallback(function () {
-            $widgetClass = call_user_func_array([new WidgetHelper(new View), 'widget'], func_get_args());
+            $widgetClass = call_user_func_array([new WidgetHelper($this->Widget->getView()), 'widget'], func_get_args());
 
             $widgetClass->Instagram = $this->getMockBuilder(get_class($widgetClass->Instagram))
                 ->setMethods(['getRecentResponse'])
                 ->getMock();
 
-            $returnValue = file_get_contents(TEST_APP . 'examples' . DS . 'recent.json');
-
-            if (in_array($this->getName(), ['testLatestNoPhotos', 'testRandomNoPhotos'])) {
-                $returnValue = json_encode(['data' => []]);
-            }
-
             $widgetClass->Instagram->method('getRecentResponse')
-                ->will($this->returnValue($returnValue));
+                ->will($this->returnCallback(function () {
+                    if (in_array($this->getName(), ['testLatestNoPhotos', 'testRandomNoPhotos'])) {
+                        return json_encode(['data' => []]);
+                    }
+
+                    return file_get_contents(TEST_APP . 'examples' . DS . 'recent.json');
+                }));
 
             return $widgetClass;
         }));
+
+        parent::setUp();
+    }
+
+    /**
+     * Called after every test method
+     * @return void
+     */
+    public function tearDown()
+    {
+        parent::tearDown();
+
+        Cache::clearAll();
     }
 
     /**
@@ -70,9 +72,8 @@ class PhotosWidgetsCellTest extends TestCase
      */
     public function testLatest()
     {
-        $widget = ME_CMS_INSTAGRAM . '.Photos::latest';
+        $widget = 'MeCmsInstagram.Photos::latest';
 
-        $result = $this->Widget->widget($widget)->render();
         $expected = [
             ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
@@ -85,9 +86,8 @@ class PhotosWidgetsCellTest extends TestCase
             '/div',
             '/div',
         ];
-        $this->assertHtml($expected, $result);
+        $this->assertHtml($expected, $this->Widget->widget($widget)->render());
 
-        $result = $this->Widget->widget($widget, ['limit' => 2])->render();
         $expected = [
             ['div' => ['class' => 'widget mb-4']],
             'h4' => ['class' => 'widget-title'],
@@ -103,19 +103,16 @@ class PhotosWidgetsCellTest extends TestCase
             '/div',
             '/div',
         ];
-        $this->assertHtml($expected, $result);
-
-        //Sets `Instagram` controller
-        $widget = $this->Widget->widget(ME_CMS_INSTAGRAM . '.Photos::latest');
-        $widget->request = $widget->request->withParam('controller', 'Instagram');
-        $this->assertEmpty($widget->render());
+        $this->assertHtml($expected, $this->Widget->widget($widget, ['limit' => 2])->render());
 
         //Tests cache
-        $fromCache = Cache::read('widget_latest_1', 'instagram');
-        $this->assertEquals(1, count($fromCache));
+        $this->assertEquals(1, count(Cache::read('widget_latest_1', 'instagram')));
+        $this->assertEquals(2, count(Cache::read('widget_latest_2', 'instagram')));
 
-        $fromCache = Cache::read('widget_latest_2', 'instagram');
-        $this->assertEquals(2, count($fromCache));
+        //Sets `Instagram` controller
+        $request = $this->Widget->getView()->getRequest()->withParam('controller', 'Instagram');
+        $this->Widget->getView()->setRequest($request);
+        $this->assertEmpty($this->Widget->widget($widget)->render());
     }
 
     /**
@@ -124,7 +121,7 @@ class PhotosWidgetsCellTest extends TestCase
      */
     public function testLatestNoPhotos()
     {
-        $this->assertEmpty($this->Widget->widget(ME_CMS_INSTAGRAM . '.Photos::latest')->render());
+        $this->assertEmpty($this->Widget->widget('MeCmsInstagram.Photos::latest')->render());
     }
 
     /**
@@ -133,7 +130,7 @@ class PhotosWidgetsCellTest extends TestCase
      */
     public function testRandom()
     {
-        $widget = ME_CMS_INSTAGRAM . '.Photos::random';
+        $widget = 'MeCmsInstagram.Photos::random';
 
         $result = $this->Widget->widget($widget)->render();
         $expected = [
@@ -168,14 +165,13 @@ class PhotosWidgetsCellTest extends TestCase
         ];
         $this->assertHtml($expected, $result);
 
-        //Sets `Instagram` controller
-        $widget = $this->Widget->widget(ME_CMS_INSTAGRAM . '.Photos::random');
-        $widget->request = $widget->request->withParam('controller', 'Instagram');
-        $this->assertEmpty($widget->render());
-
         //Tests cache
-        $fromCache = Cache::read('widget_latest_12', 'instagram');
-        $this->assertEquals(12, count($fromCache));
+        $this->assertEquals(12, count(Cache::read('widget_latest_12', 'instagram')));
+
+        //Sets `Instagram` controller
+        $request = $this->Widget->getView()->getRequest()->withParam('controller', 'Instagram');
+        $this->Widget->getView()->setRequest($request);
+        $this->assertEmpty($this->Widget->widget($widget)->render());
     }
 
     /**
@@ -184,6 +180,6 @@ class PhotosWidgetsCellTest extends TestCase
      */
     public function testRandomNoPhotos()
     {
-        $this->assertEmpty($this->Widget->widget(ME_CMS_INSTAGRAM . '.Photos::random')->render());
+        $this->assertEmpty($this->Widget->widget('MeCmsInstagram.Photos::random')->render());
     }
 }